requirements.txt
beautifulsoup4
dash
dash-core-components
dash-defer-js-import
dash-html-components
dash-renderer
dash-table
flake8
Flask
Flask-Compress
folium
pandas
pandocfilters
path
pathlib2
pathtools
Pillow
seaborn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dash_bootstrap_components
#torch==1.5.1+cpu
#torchvision==0.6.1+cpu